pyproject.toml
[project]
name = "soupawhisper"
version = "2.0.0"
description = "Voice dictation tool using Groq Whisper API"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "pynput>=1.8.1",
    "requests>=2.32.5",
    "evdev>=1.7.0; sys_platform == 'linux'",
    "flet>=0.25.0",
    "pystray>=0.19.0",
    "pillow>=10.0.0",
    "pyperclip>=1.8.0; sys_platform == 'win32'",
]

[project.optional-dependencies]
tray = [
    "pygobject>=3.50.0; sys_platform == 'linux'",
]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
]

[project.scripts]
soupawhisper = "soupawhisper.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/soupawhisper"]

[tool.hatch.build.targets.wheel.force-include]
"src/soupawhisper/gui/assets" = "soupawhisper/gui/assets"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"

[dependency-groups]
dev = [
    "flet-web>=0.80.1",
    "playwright>=1.57.0",
    "pytest-playwright>=0.7.2",
    "uvicorn>=0.40.0",
]
